pytest: Separate Rapid3 tests
